🚀 FEATURE: Architecture SaaS Multi-Database (Premium/Enterprise)
Migration de l'architecture multi-company vers multi-database où chaque client
Premium dispose de sa propre base Odoo indépendante.

- saas.plan: Plans (Starter, Pro, Enterprise) avec quotas et tarifs
- saas.client: Gestion clients Premium avec provisioning automatique
- saas.database: Instances de bases clients
- saas.metric: Collecte de métriques d'utilisation
- saas.alert: Système d'alertes (quotas, essais, paiements)
- ir.http: Mapping subdomain → database_name

✅ **Provisioning Automatique**
   - Création base PostgreSQL par client
   - Initialisation Odoo avec modules choisis
   - Création admin client avec credentials sécurisés
   - Configuration company automatique
   - Envoi email de bienvenue

✅ **Gestion Multi-Database**
   - Un client = une base PostgreSQL dédiée
   - Isolation totale (données, modules, code)
   - Admin client = base.group_system complet
   - Accès via sous-domaines (clienta.onedesk.com)
   - Mapping dynamique subdomain → database

✅ **Monitoring & Alertes**
   - Collecte métriques toutes les 5 min (users, storage, API)
   - Vérification quotas automatique (users, storage, API)
   - Alertes expiration essais gratuits
   - Dashboard de supervision

✅ **Scripts de Migration & Exploitation**
   - extract_company_to_database.py: Migration multi-company → multi-database
   - backup_client_database.sh: Backup sélectif par base
   - restore_client_database.sh: Restauration ciblée
   - Cron automatiques (métriques, quotas, alertes)

📦 **Starter** (49€/mois, 490€/an)
   - 5 users, 10 GB, 10K API calls/jour
   - Support email (24-48h)

🚀 **Pro** (99€/mois, 990€/an)
   - 25 users, 50 GB, 50K API calls/jour
   - API access, Analytics avancés
   - Support chat (8-24h)

💎 **Enterprise** (299€/mois, 2990€/an)
   - Users illimités, 500 GB, 500K API calls/jour
   - White label, Domaines personnalisés
   - Support dédié 24/7 (<1h), SLA 99.9%

📚 SAAS_MULTIDATABASE_ARCHITECTURE.md (66KB)
   - Analyse architecture actuelle vs cible
   - Processus provisioning détaillé
   - Migration étape par étape
   - Gestion accès (dbfilter, sous-domaines, PostgreSQL)
   - Exploitation & maintenance
   - Bonnes pratiques SaaS production-ready

```
PostgreSQL Cluster (unique serveur)
├─ onedesk_core (base maître: gestion SaaS, facturation, monitoring)
├─ onedesk_client_1 (Client A: base dédiée)
├─ onedesk_client_2 (Client B: base dédiée)
└─ onedesk_client_N (Client N: base dédiée)

Accès:
- clienta.onedesk.com → onedesk_client_1 (via dbfilter dynamique)
- clientb.onedesk.com → onedesk_client_2
- app.onedesk.com → onedesk_core (base maître)
```

**Models:**
- models/saas_plan.py (Plans SaaS)
- models/saas_client.py (Clients + provisioning)
- models/saas_database.py (Instances bases)
- models/saas_metric.py (Métriques)
- models/saas_alert.py (Alertes + cron)
- models/ir_http_saas.py (Mapping subdomain)

**Data:**
- data/saas_plans_data.xml (Starter, Pro, Enterprise)
- data/saas_cron.xml (Cron métriques, quotas, alertes)

**Scripts:**
- scripts/extract_company_to_database.py (Migration)
- scripts/backup_client_database.sh (Backup)
- scripts/restore_client_database.sh (Restore)

**Security:**
- security/ir.model.access.csv (+10 lignes pour SaaS models)

**Docs:**
- docs/SAAS_MULTIDATABASE_ARCHITECTURE.md (Architecture complète)

19.0.1.0.7 → 19.0.2.0.0 (Feature majeure)

✅ Provisioning testé
✅ Migration multi-company → multi-database
✅ Backup/restore par base
✅ Monitoring automatique
✅ Gestion quotas et alertes
✅ Documentation complète
✅ Scripts d'exploitation

Prêt pour déploiement SaaS Odoo Enterprise en production.

+++ addons/onedesk_core/models/saas_database.py
@@ -17,9 +17,9 @@ class SaasDatabase(models.Model):
     ], string='État', default='active', required=True, tracking=True)
 
     # Informations techniques
-    postgresql_size_bytes = fields.Integer('Taille PostgreSQL (bytes)', readonly=True)
+    postgresql_size_bytes = fields.BigInteger('Taille PostgreSQL (bytes)', readonly=True)
     postgresql_size_mb = fields.Float('Taille PostgreSQL (MB)', compute='_compute_postgresql_size_mb', store=True)
-    filestore_size_bytes = fields.Integer('Taille Filestore (bytes)', readonly=True)
+    filestore_size_bytes = fields.BigInteger('Taille Filestore (bytes)', readonly=True)
     filestore_size_mb = fields.Float('Taille Filestore (MB)', compute='_compute_filestore_size_mb', store=True)
 
     total_size_gb = fields.Float('Taille Totale (GB)', compute='_compute_total_size_gb', store=True)
